Devstart now builds dev and deployed docs
"npm run devstart" now uses swagger-autogen to generate two swagger files: one for when running on localhost and one for when running deployed. Environmental variables determines which gets used. Pretty inefficient, but I wanted to see if I could find a way to make it so I don't have to manually run  separate scripts or modify files to update and test my documentation.